requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
requests
python-dotenv
